small cleanup/refactor, delete unused code. put internal plugins in same file
