+++ src/NiceHashMiner/Forms/Components/AlgorithmSettingsControl.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Windows.Forms;
 using NiceHashMiner.Algorithms;
+using NiceHashMiner.Utils;
 
 namespace NiceHashMiner.Forms.Components
 {
