+++ src/NiceHashMiner/Forms/LoginForm.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using NiceHashMiner.Utils;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
