--- src/NiceHashMiner/Miners/IntegratedPlugins/BMinerIntegrated/BMinerIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class BMinerIntegratedPlugin : BMiner.BMinerPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "BMiner";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
