--- src/NiceHashMiner/Miners/IntegratedPlugins/ClaymoreDual14Integrated/ClaymoreDual14IntegratedPlugin.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class ClaymoreDual14IntegratedPlugin : ClaymoreDual14.ClaymoreDual14Plugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "ClaymoreDual14+";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
