--- src/NiceHashMiner/Miners/IntegratedPlugins/EWBFIntegrated/EWBFIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class EWBFIntegratedPlugin : EWBF.EwbfPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "Ewbf";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
