--- src/NiceHashMiner/Miners/IntegratedPlugins/EthlargementIntegrated/EthlargementIntegratedPlugin.cs
@@ -1,24 +1,0 @@
-﻿using NHM.Common;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class EthlargementIntegratedPlugin : Ethlargement.Ethlargement, IntegratedPlugin
-    {
-        public static EthlargementIntegratedPlugin Instance { get; } = new EthlargementIntegratedPlugin();
-        EthlargementIntegratedPlugin() : base("Ethlargement")
-        { }
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
