--- src/NiceHashMiner/Miners/IntegratedPlugins/GMinerIntegrated/GMinerIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class GMinerIntegratedPlugin : GMinerPlugin.GMinerPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "GMiner";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
