--- src/NiceHashMiner/Miners/IntegratedPlugins/NBMinerIntegrated/NBMinerIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class NBMinerIntegratedPlugin : NBMiner.NBMinerPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "NBMiner";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
