--- src/NiceHashMiner/Miners/IntegratedPlugins/PhoenixIntegrated/PhoenixIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class PhoenixIntegratedPlugin : Phoenix.PhoenixPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "Phoenix";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
