+++ src/NiceHashMiner/Miners/IntegratedPlugins/SGminerIntegratedHub/SGminerAvemoreIntegratedPlugin.cs
@@ -1,5 +1,4 @@
-﻿using MinerPlugin;
-using MinerPluginToolkitV1;
+﻿using MinerPluginToolkitV1;
 using MinerPluginToolkitV1.Configs;
 using MinerPluginToolkitV1.Interfaces;
 using NHM.Common.Algorithm;
@@ -8,8 +7,6 @@ using NHM.Common.Enums;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NiceHashMiner.Miners.IntegratedPlugins
 {
