+++ src/NiceHashMiner/Miners/IntegratedPlugins/SGminerIntegratedHub/SGminerGMIntegratedPlugin.cs
@@ -1,14 +1,11 @@
-﻿using MinerPlugin;
-using MinerPluginToolkitV1;
+﻿using MinerPluginToolkitV1;
 using MinerPluginToolkitV1.Configs;
 using MinerPluginToolkitV1.Interfaces;
-using NHM.Common;
 using NHM.Common.Algorithm;
 using NHM.Common.Device;
 using NHM.Common.Enums;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 
 namespace NiceHashMiner.Miners.IntegratedPlugins
