--- src/NiceHashMiner/Miners/IntegratedPlugins/TTMinerIntegrated/TTMinerIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    public class TTMinerIntegratedPlugin : TTMiner.TTMinerPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "TTMiner";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
