+++ src/NiceHashMiner/Miners/IntegratedPlugins/VC_REDIST_x64_2015_DEPENDENCY_PLUGIN.cs
@@ -8,9 +8,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace NiceHashMiner.Miners.IntegratedPlugins
 {
@@ -59,11 +56,6 @@ namespace NiceHashMiner.Miners.IntegratedPlugins
         public IEnumerable<string> CheckBinaryPackageMissingFiles()
         {
             return BinaryPackageMissingFilesCheckerHelpers.ReturnMissingFiles("", new List<string> { VcRedistBinPath() });
-        }
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
         }
 
         public void InstallVcRedist()
