--- src/NiceHashMiner/Miners/IntegratedPlugins/WildRigIntegrated/WildRigIntegratedPlugin.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    public class WildRigIntegratedPlugin : WildRig.WildRigPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "WildRig";
-
-        public bool Is3rdParty => true;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
