--- src/NiceHashMiner/Miners/IntegratedPlugins/XmrStakIntegrated/XmrStakIntegratedPlugin.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace NiceHashMiner.Miners.IntegratedPlugins
-{
-    class XmrStakIntegratedPlugin : XmrStak.XmrStakPlugin, IntegratedPlugin
-    {
-        public override string PluginUUID => "XmrStak";
-
-        public bool Is3rdParty => false;
-
-        IEnumerable<string> IntegratedPlugin.GetMinerBinsUrls()
-        {
-            return MinersBinsUrls.GetMinerBinsUrlsForPlugin(PluginUUID);
-        }
-    }
-}
