+++ src/NiceHashMiner/Miners/Miner.cs
@@ -1,5 +1,4 @@
-﻿using NiceHashMiner.Miners.Grouping;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
