+++ src/NiceHashMiner/Miners/MiningManager.cs
@@ -1,20 +1,15 @@
 ﻿using NiceHashMiner.Configs;
 using NiceHashMiner.Devices;
-using NiceHashMiner.Interfaces;
 using NiceHashMiner.Miners.Grouping;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using System.Timers;
-using System.Windows.Forms;
 using NiceHashMiner.Algorithms;
-using NiceHashMiner.Benchmarking;
 using NiceHashMiner.Stats;
 using NiceHashMiner.Switching;
 using NHM.Common.Enums;
-using Timer = System.Timers.Timer;
 using NHM.Common;
 using NiceHashMiner.Miners.IntegratedPlugins;
 using MinerPlugin;
